Reorganized plugin code into a module
