+++ qgis_earthengine_plugin/Map.py
