+++ qgis_earthengine_plugin/metadata.txt
