+++ test/test_init.py
@@ -47,7 +47,12 @@ class TestInit(unittest.TestCase):
         ]
 
         file_path = os.path.abspath(
-            os.path.join(os.path.dirname(__file__), os.pardir, "metadata.txt")
+            os.path.join(
+                os.path.dirname(__file__),
+                os.pardir,
+                "qgis_earthengine_plugin",
+                "metadata.txt",
+            )
         )
         LOGGER.info(file_path)
         metadata = []
